moves types to types.ts
