+++ utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./merkleTree";
 export * from "./snapshotHub";
+export * from "./types";
